Experimental Cleanup tweak
DO NOT BASE ANY CODE ON THIS COMMIT

This commit is experimental to see if CI integration tests can run faster with
async(false) API calls in the CloudFoundryCleaner code.

It is arguable that this behaviour is actually better behaved and less
disruptive (to network connections, for example) than the corresponding
async(true) cases, which causes tasks to be scheduled and tracked, polled for
and eventually discarded.

async(false), by contrast holds up a network connection (port) and could
therefore easily swamp the pool of usable ports on the CloudController side.
On the other hand, the reactor streams gating of dispatchable tasks means that
only a limited number of these could possibly be running simultaneously from
each client, providing a natural throttle on this effect.