Trial replacement of Relationship class by Id class
Id.of(x) replaces Relationship.builder(x).id().build() and the check on x
being non-null is made immediately, not deferred until we look at the data.

1. The name "Relationship" is misleading: this object is just the image of a
relationship, and there is confusion with the method Request.relationship(a,b)
where "relationships" is a @Singular map on a request.

2. The builder pattern is overkill for an object with no optional parts and a
single mandatory part which must not be null.

3. An application (or layer implementation) only ever creates these for use in
a request. There is no need of one containing null, nor of passing or using the
builder() of one. Indeed, it risks obscure bugs if it happens.

4. The tests for validity of these objects are easy to forget, and normally
occur far too late in the API for the caller to do anything sensible about it.
Better to throw an exception early (in an application or implementation layer)
than fail a sequence of CF calls late.

5. It is never the case that CF responds with a null guid value.  If it is
there, it is non-null. Moreover you cannot signal a null value in a reactor-
stream, so nulls will often throw exceptions before getting to API checking
code anyway.

6. `Relationship` should, in any case, be annotated with @Data so as to
support test comparisons and implicit toString() operations. In such cases the
name Relationship is cluttering and confusing.

NB: The missing checks on Relationship(s) fields were missing because one has
to explicitly check them, and no-one remembered to write a test for this.
(What test would spot it when you've forgotten it already?) With this change
you never have to test for it; it is always explicitly checked —— just like
Optional.of(x) does. The "transitive" validity checks are unnecessary.